This patch syncs the gpu dev branch with trunk.  It also fixes a handful of bugs to get MPI working again.  This patch is full of debugger print messages, I'm just committing it to get the bug fixes in.  I'll be removing those messages shortly
Former-commit-id: 11c735138ffad94f6dbb471fa03e8c0fe234ee62

+++ gpu_dev/src/CCA/Components/MPM/ParticleCreator/ImplicitParticleCreator.cc
@@ -58,7 +58,7 @@ ImplicitParticleCreator::initializeParticle(const Patch* patch,
                                             MPMMaterial* matl,
                                             Point p, IntVector cell_idx,
                                             particleIndex i,
-                                            CCVariable<short int>& cellNAPI,
+                                            CCVariable<int>& cellNAPI,
                                             ParticleVars& pvars)
 {
